Turned all macros to functions. Marked everything as G_GNUC_PRINTF and
	* include/glibtop/error.h:
	* lib/parameter.c: (glibtop_set_parameter_l):
	* lib/read_data.c: (glibtop_read_data_l):
	* sysdeps/common/error.c: (print_server_name), (glibtop_error_vr),
	(glibtop_error_io_vr), (glibtop_warn_vr), (glibtop_warn_io_vr),
	(glibtop_error_r), (glibtop_warn_r), (glibtop_error_io_r),
	(glibtop_warn_io_r), (glibtop_error), (glibtop_warn),
	(glibtop_error_io), (glibtop_warn_io): Turned all macros to functions.
	Marked everything as G_GNUC_PRINTF and fixed format bugs.

	* sysdeps/common/fsusage.c: Marked some functions as G_GNUC_INTERNAL

+++ lib/parameter.c
@@ -32,9 +32,9 @@
 
 #define _check_data(size)	\
 	if ((data_ptr == NULL) || (data_size != size)) { \
-		glibtop_warn_r (server, "glibtop_set_parameter (%d): " \
+		glibtop_warn_r (server, "glibtop_set_parameter (%u): " \
 				"Expected %lu bytes but got %lu.", \
-				parameter, size, data_size); \
+				(unsigned) parameter, (unsigned long) size, (unsigned long) data_size); \
 		return; \
 	}
 
@@ -82,7 +82,7 @@ glibtop_set_parameter_l (glibtop *server, const unsigned parameter,
 		break;
 	case GLIBTOP_PARAM_FEATURES:
 		/* You should not be allowed to set this field. */
-		glibtop_warn_r (server, "glibtop_set_parameter (%d): " \
+		glibtop_warn_r (server, "glibtop_set_parameter (%u): " \
 				"Cannot modify read-only value.",
 				parameter);
 		break;
